requirements.txt
aws-cdk-lib>=2.67.0
constructs>=10.0.0,<11.0.0
aws_solutions_constructs.aws-wafwebacl-cloudfront>=2.33.0
cdk-nag>=2.10.0